requirements.txt
django==4.2.1
Pillow==9.2.0
djangorestframework
psycopg2
uwsgi
